[draft 1] Implementation of sub-second accounting timer (currently 100ms precision). Cost per iteration order of 1-10 us.

+++ src/common/slurm_acct_gather_infiniband.c
@@ -72,7 +72,7 @@ static plugin_context_t *g_context = NULL;
 static pthread_mutex_t g_context_lock =	PTHREAD_MUTEX_INITIALIZER;
 static bool init_run = false;
 static bool acct_shutdown = true;
-static int freq = 0;
+static float freq = 0.0;
 
 static void *_watch_node(void *arg)
 {
@@ -141,7 +141,7 @@ extern int acct_gather_infiniband_fini(void)
 	return rc;
 }
 
-extern int acct_gather_infiniband_startpoll(uint32_t frequency)
+extern int acct_gather_infiniband_startpoll(float frequency)
 {
 	int retval = SLURM_SUCCESS;
 	pthread_attr_t attr;
@@ -160,7 +160,7 @@ extern int acct_gather_infiniband_startpoll(uint32_t frequency)
 
 	freq = frequency;
 
-	if (frequency == 0) {   /* don't want dynamic monitoring? */
+	if (frequency == 0.0) {   /* don't want dynamic monitoring? */
 		debug2("acct_gather_infiniband dynamic logging disabled");
 		return retval;
 	}
