+++ src/common/slurm_acct_gather_infiniband.h
@@ -74,7 +74,7 @@ typedef struct acct_network_data {
 
 extern int acct_gather_infiniband_init(void); /* load the plugin */
 extern int acct_gather_infiniband_fini(void); /* unload the plugin */
-extern int acct_gather_infiniband_startpoll(uint32_t frequency);
+extern int acct_gather_infiniband_startpoll(float frequency);
 
 extern int acct_gather_infiniband_g_node_update(void);
 /*
